add dialog

+++ lib/main.dart
@@ -66,7 +66,30 @@ class _HomePageState extends State<HomePage> {
                     Text(memos[index], style: TextStyle(fontSize: 24)),
                     IconButton(
                       onPressed: () {
-                        deleteMemo(index);
+                        showDialog(
+                          context: context,
+                          builder: (context) {
+                            return AlertDialog(
+                              title: Text('メモを削除'),
+                              content: Text('本当に削除しますか？この作業は取り消しできません。'),
+                              actions: [
+                                TextButton(
+                                  onPressed: () {
+                                    Navigator.of(context).pop();
+                                  },
+                                  child: Text('キャンセル'),
+                                ),
+                                TextButton(
+                                  onPressed: () {
+                                    deleteMemo(index);
+                                    Navigator.of(context).pop();
+                                  },
+                                  child: Text('削除'),
+                                ),
+                              ],
+                            );
+                          },
+                        );
                       },
                       icon: Icon(Icons.delete),
                     ),
